requirements.txt
annotated-types==0.7.0
certifi==2024.7.4
charset-normalizer==3.3.2
colorama==0.4.6
customtkinter==5.2.2
darkdetect==0.8.0
engineering-notation==0.10.0
idna==3.7
loguru==0.7.2
numpy==2.0.1
packaging==24.1
pandas==2.2.2
pillow==10.4.0
pydantic==2.8.2
pydantic_core==2.20.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
six==1.16.0
tk-tools==0.16.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
win32-setctime==1.1.0
